Documentation updates and bug fixes
Added some more detailed documentation on validators and plugins. Plugin
runs are now also enclosed in a try block. The error_mode setting is
used do determine how to handle a plugin error. With error_mode set to
break, the plugin exception is raised and the test will end. Otherwise,
the plugin exception will cancel the current tester, but continue with
the next one.