Remove failed stop actions from atexit
If a plugin raises an exception during the stop action, it is not set to
stopped. This means that the stop action runs again during the program
exit, which probably causes the same exception again, without any
handling.

This commit removes plugins with failed stop actions from the exit
events.