[tests] Fix status codes for conditional tests
The conditional tests that cause errors expected a status code of 9
(Tricot-Exception), which was actually wrong all the time. Due to a bug
within the test execution the exceptions were caught in a wrong way and
the false error code was returned. We fixed this in one of the latest
commits, but as the tests still expected the wrong status code, they
failed.

This should now be fixed.