Add --skip-until and --continue-from options
Added the --skip-until and --continue-from options. Both can be used to
skip tests until a certain test / tester id is reached. This is useful
when the error mode was set to break and you want to continue the test
on the last test that produced an error.

Internally, both options are handeled in the same way. The only reason
why we introduce two options is a better user experience.