Allow lists in command variables
tricot commands are specified as list values. Variables too. So far we
converted variables to strings before usage. However, this is probably
not intended when you want to use a variable list within a command.
Instead it is more likely that you want to insert the variable list into
the command. This is the way it is now implemented.