Implemented command timeout by using process group
The current implementation of a command timeout did only work for
programms that are started dircectly (PID of resulting program is the
same as the program that was invoked to start it). When you want to test
a bash script that calls other programs, this is no longer true. This
caused issues because it was not possible to kill processes like this.

This commit implements process killing by using a process group.